+++ evirdle/script.js
@@ -779,6 +779,7 @@ const msOffset = Date.now() - offsetFromDate
 const dayOffset = msOffset / 1000 / 60 / 60 / 24
 const queryParams = new URLSearchParams(window.location.search)
 let easyMode = false
+let gameOver = false
 
 // Use as answers only those words with no doublets (e.g. not FIZZY or BLOOM).
 let currentWords = targetWords.filter(w => !w.match("(.)\\1"))
@@ -800,23 +801,26 @@ function stopInteraction() {
 }
 
 function handleMouseClick(e) {
-  if (e.target.matches("[data-key]")) {
+  if (e.target.matches("[data-key]") && !gameOver) {
     pressKey(e.target.dataset.key)
-  } else if (e.target.matches("[data-enter]")) {
+  } else if (e.target.matches("[data-enter]") && !gameOver) {
     submitGuess()
-  } else if (e.target.matches("[data-delete]")) {
+  } else if (e.target.matches("[data-delete]") && !gameOver) {
     deleteKey()
   } else if (e.target.matches("#help-button")) {
     document.getElementById("help-modal").hidden = false
-  } else if (e.target.matches("#close-help-button")) {
-    document.getElementById("help-modal").hidden = true
   } else if (e.target.matches("#settings-button")) {
     document.getElementById("settings-modal").hidden = false
-  } else if (e.target.matches("#close-settings-button")) {
+  } else if (e.target.matches(".close-modal-button") || e.target.matches(".overlay")) {
+    document.getElementById("help-modal").hidden = true
     document.getElementById("settings-modal").hidden = true
   } else if (e.target.matches("#easy-mode")) {
-    window.easyMode = e.target.toggleAttribute("checked")
-    saveSettings()
+    if (e.target.matches(".disabled-control")) {
+      showAlert("Start a new game first")
+    } else {
+      window.easyMode = e.target.toggleAttribute("checked")
+      saveSettings()
+    }
   } else if (e.target.matches("#dark-theme")) {
     const on = e.target.toggleAttribute("checked")
     document.querySelector("body").classList.toggle("nightmode", on)
@@ -825,6 +829,16 @@ function handleMouseClick(e) {
     const on = e.target.toggleAttribute("checked")
     document.querySelector("body").classList.toggle("colorblind", on)
     saveSettings()
+  } else if (e.target.matches("#share-button")) {
+    if (navigator.clipboard.writeText) {
+      navigator.clipboard.writeText(createUnicodeGameTranscript()).then(function() {
+        showAlert("Copied to clipboard")
+      }, function (err) {
+        showAlert("Copying failed: " + e)
+      })
+    } else {
+      showAlert("Copying failed: navigator.clipboard.writeText is not supported")
+    }
   }
 }
 
@@ -851,17 +865,10 @@ function restoreSettings() {
 function handleKeyPress(e) {
   if (e.key === "Enter") {
     submitGuess()
-    return
-  }
-
-  if (e.key === "Backspace" || e.key === "Delete") {
+  } else if (e.key === "Backspace" || e.key === "Delete") {
     deleteKey()
-    return
-  }
-
-  if (e.key.match(/^[a-z]$/)) {
+  } else if (e.key.match(/^[a-z]$/)) {
     pressKey(e.key)
-    return
   }
 }
 
@@ -902,6 +909,7 @@ function submitGuess() {
   }
 
   stopInteraction()
+  document.querySelector("#easy-mode").classList.toggle("disabled-control", true)
 
   let results = {}
   for (let i = 0; i < currentWords.length; ++i) {
@@ -1038,22 +1046,52 @@ function shakeTiles(tiles) {
   })
 }
 
+function createUnicodeGameTranscript() {
+  console.assert(window.gameOver)
+  const usedRows = guessGrid.querySelectorAll("[data-letter]").length / WORD_LENGTH
+
+  let transcript = ""
+  let accumulator = []
+  const greenSquare = document.querySelector("body").classList.contains("colorblind") ? "\u{1F7E7}" : "\u{1F7E9}"
+  const yellowSquare = document.querySelector("body").classList.contains("colorblind") ? "\u{1F7E6}" : "\u{1F7E8}"
+  const whiteSquare = document.querySelector("body").classList.contains("nightmode") ? "\u{2B1B}" : "\u{2B1C}"
+  for (let e of guessGrid.querySelectorAll("[data-letter]")) {
+    accumulator.push(e.dataset.state == "correct" ? greenSquare :
+                     e.dataset.state == "present" ? yellowSquare : whiteSquare)
+    if (accumulator.length === WORD_LENGTH) {
+      transcript += "\n" + accumulator.join("")
+      accumulator = []
+    }
+  }
+  console.assert(accumulator.length === 0)
+
+  if (transcript.endsWith(greenSquare+greenSquare+greenSquare+greenSquare+greenSquare)) {
+    return (window.easyMode ? "Easy " : "") + "Evirdle " + currentWords[0].toUpperCase() + " " + usedRows + "/6\n" + transcript
+  } else {
+    return (window.easyMode ? "Easy " : "") + "Evirdle " + currentWords[0].toUpperCase() + " X/6\n" + transcript
+  }
+}
+
 function checkWinLose(youWon, tiles) {
   const usedRows = guessGrid.querySelectorAll("[data-letter]").length / WORD_LENGTH
   const remainingRows = guessGrid.querySelectorAll(":not([data-letter])").length / WORD_LENGTH
 
   if (youWon) {
+    stopInteraction()
     const compliments = ["Genius", "Magnificent", "Impressive", "Splendid", "Great", "Phew"]
     showAlert(compliments[usedRows - 1], 5000)
     danceTiles(tiles)
+  } else if (remainingRows === 0) {
     stopInteraction()
+    showAlert(currentWords[0].toUpperCase(), null)
+  } else {
     return
   }
 
-  if (remainingRows === 0) {
-    showAlert(currentWords[0].toUpperCase(), null)
-    stopInteraction()
-  }
+  let e = document.querySelector("#share-button")
+  e.classList.toggle("disabled-button", false)
+  window.gameOver = true
+  startInteraction()
 }
 
 function danceTiles(tiles) {
